requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
apscheduler==3.10.4
sqlalchemy==2.0.36
requests==2.32.3
httpx==0.27.2
python-dotenv==1.0.1
gunicorn==22.0.0
pillow==10.4.0
tweepy==4.14.0
werkzeug==3.0.3
python-telegram-bot==21.6
